Add new factory and seeds, learn how to refresh migration and reload whole project

+++ database/migrations/2025_03_14_095322_add_fields_to_listings_table.php
@@ -12,9 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('listings', function (Blueprint $table) {
-            $table->unsignedInteger('beds');
-            $table->unsignedInteger('baths');
-            $table->unsignedInteger('area');
+            $table->unsignedTinyInteger('beds');
+            $table->unsignedTinyInteger('baths');
+            $table->unsignedSmallInteger('area');
 
             $table->tinyText('city');
             $table->tinyText('code');
